audio changes

+++ AuralNano/main.py
@@ -34,10 +34,12 @@ while True:
         continue
 
     if "new music" in speech:
-        os.remove("received_midi_file.mid")
+        if os.path.exists("received_midi_file.mid"):
+            os.remove("received_midi_file.mid")
+
         playText("Upload a clear screenshot of your sheet music on the website.")
         socket.send_multipart([client_id, b"NEW_MUSIC"])
-        playText("I am processing your sheet music is processing now. This may take a while, so feel free to practice on your own until you hear me tell you I'm ready")
+        playText("I will be processing your sheet music. This may take a while, so feel free to practice on your own until you hear me tell you I'm ready")
         receive_file_and_string(socket, "received_midi_file.mid")
         playText("I have finished processing the sheet music. Feel free to say any of the commands you see on the screen")
 
@@ -48,8 +50,9 @@ while True:
         testName = auralGlobals.sheetMusicName + "-" + datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
         socket.send_multipart([client_id, b"TEST", testName.encode('utf-8')])
         grade, feedback = testMusic()
-        playText(f"Music Test Completed: You played {auralGlobals.sheetMusicName} and receieved a music grade of {grade}%. Check your test cards on the website for more details and feedback.")
         socket.send_multipart([client_id, b"TEST_DONE"])
+        playText(f"Music Test Completed: You played {auralGlobals.sheetMusicName} and receieved a music grade of {grade}%. Check your test cards on the website for more details and feedback.")
+        
         print("Test results sent to Visual Nano")
 
     elif "play" in speech:
